+++ Editor/QuickJSON.cs
@@ -38,8 +38,15 @@ namespace Reallusion.Import
             text = jsonText;
             index = startIndex;
             values = new List<MultiValue>();
-            if (index == 0) Brace();
-            Parse();
+            try
+            {
+                if (index == 0) Brace();
+                Parse();
+            }
+            catch
+            {
+                Debug.LogError("Unable to Parse JSON text...");
+            }
             text = null;
         }
 
@@ -106,9 +113,16 @@ namespace Reallusion.Import
         {
             string value = Value();
 
-            if (int.TryParse(value, out int tryInt))
+            if (int.TryParse(value,
+                    System.Globalization.NumberStyles.Integer, 
+                    System.Globalization.CultureInfo.InvariantCulture, 
+                    out int tryInt))
                 values.Add(new MultiValue(name, tryInt));
-            else if (float.TryParse(value, out float tryFloat))
+
+            else if (float.TryParse(value, 
+                    System.Globalization.NumberStyles.Number, 
+                    System.Globalization.CultureInfo.InvariantCulture, 
+                    out float tryFloat))
                 values.Add(new MultiValue(name, tryFloat));
         }
 
@@ -199,9 +213,9 @@ namespace Reallusion.Import
         public MultiValue GetValue(string name)
         {
             foreach (MultiValue mv in values)
-                if (mv.Key == name) return mv;            
+                if (mv.Key.iEquals(name)) return mv;            
 
-            return new MultiValue("", false);
+            return new MultiValue("None");
         }
 
         public MultiValue GetValue(int index)
@@ -209,7 +223,7 @@ namespace Reallusion.Import
             if (values.Count > index)
                 return values[index];            
 
-            return new MultiValue("", false);
+            return new MultiValue("None");
         }
 
         public QuickJSON FindObjectWithKey(string keySearch)
@@ -233,6 +247,27 @@ namespace Reallusion.Import
             return null;
         }
 
+        public bool PathExists(string path)
+        {
+            string[] paths = path.Split('/');
+
+            return PathExists(paths);
+        }
+
+        public bool PathExists(string[] paths)
+        {
+            if (paths.Length > 0)
+            {
+                MultiValue mv = GetValue(paths[0]);
+                if (paths.Length > 1 && mv.Type == MultiType.Object)
+                    return mv.ObjectValue.PathExists(paths.Skip(1).ToArray());
+                else if (mv.Type != MultiType.None)
+                    return true;
+            }            
+
+            return false;
+        }
+
         public QuickJSON GetObjectAtPath(string path)
         {
             string[] paths = path.Split('/');
@@ -249,7 +284,7 @@ namespace Reallusion.Import
                     return mv.ObjectValue.GetObjectAtPath(paths.Skip(1).ToArray());
                 else if (mv.Type == MultiType.Object)
                     return mv.ObjectValue;
-            }
+            }            
 
             return null;
         }
@@ -398,7 +433,7 @@ namespace Reallusion.Import
                     return mv.ObjectValue.Vector2Value;
             }
 
-            return Vector3.zero;
+            return Vector2.zero;
         }
 
         public Color ColorValue
@@ -477,7 +512,7 @@ namespace Reallusion.Import
 
     }
 
-    public enum MultiType { Integer, Float, Bool, String, Object }
+    public enum MultiType { Integer, Float, Bool, String, Object, None }
 
     public struct MultiValue
     {
@@ -528,6 +563,13 @@ namespace Reallusion.Import
                 return Type == MultiType.Object ? 
                     (QuickJSON)objectValue : null; 
             } 
+        }
+
+        public MultiValue(string name)
+        {
+            Key = name;
+            Type = MultiType.None;
+            objectValue = null;
         }
 
         public MultiValue(string name, bool value)