+++ CHANGELOG.md
@@ -1,9 +1,20 @@
 Changelog
 =========
 
+### TODO
+- Tiling and offset attaching baked textures?
+- Back button in the import window.
+
+### v 0.1.3
+- Normal strength settings for all relevant custom shaders.
+- All normal map, micro normal map and blend normal map textures default to Bump mode when empty.
+- Emission texture and Emissive color properties added to all custom shaders.
+- Fix localization issue with number conversion reading from Json data.
+- Logs errors when Json data fails to read.
+
 ### v 0.1.1
 - Bake button will not activate until the character has been imported and set-up with High-Quality materials.
 - Animation button will not activate until the character has been imported and set-up with either Default or High-Quality materials.
 
-### v 0.1.0 
+### v 0.1.0
 - Initial release.